IPC and Menu refactor
