+++ src/main/modules/window/abs-window-ipc.ts
@@ -1,4 +1,9 @@
-import { ipcMain, type BaseWindowConstructorOptions } from 'electron'
+import {
+    ipcMain,
+    type BaseWindowConstructorOptions,
+    type IpcMainEvent,
+} from 'electron'
+import Logger from '@main/modules/logger'
 
 import {
     Channel,
@@ -10,8 +15,10 @@ import {
 } from '@src/types'
 
 import Status from '@main/modules/store/status'
-import Bookmarks from '@main/modules/store/bookmarks'
+import Shortcut from '@main/modules/store/shortcut'
 import Anchors from '@main/modules/store/anchors'
+import Popup from '@main/modules/store/popup'
+import Bookmarks from '@main/modules/store/bookmarks'
 
 import { AbsWindowMenu } from './abs-window-menu'
 
@@ -26,10 +33,12 @@ export abstract class AbsWindowIPC extends AbsWindowMenu {
         ipcMain.on(Channel.SWITCH, this.onSwitch.bind(this))
         ipcMain.on(Channel.HISTORY, this.onHistory.bind(this))
         ipcMain.on(Channel.BOOKMARK, this.onBookmarks.bind(this))
+        ipcMain.on(Channel.ANCHOR, this.onAnchors.bind(this))
+        ipcMain.on(Channel.POPUP_BLOCKER, this.onPopupBlocker.bind(this))
     }
 
-    private onInfo(
-        _: Electron.IpcMainEvent,
+    private async onInfo(
+        _: IpcMainEvent,
         handler: RequestHandler,
         data: Partial<Info>,
     ) {
@@ -40,20 +49,31 @@ export abstract class AbsWindowIPC extends AbsWindowMenu {
 
         if (handler === RequestHandler.REQUEST) {
             if (data) {
-                this.centre.sendLocation(
-                    this.browser.webContents.getTitle(),
-                    this.browser.webContents.getURL(),
+                this.centre.webContents.send(
+                    Channel.INFO,
+                    RequestHandler.RESPONSE,
+                    {
+                        title: this.browser.webContents.getTitle(),
+                        url: this.browser.webContents.getURL(),
+                    },
                 )
                 return
             }
-            this.centre.sendInfo(
-                this.browser.webContents.session.getCacheSize(),
+
+            this.centre.webContents.send(
+                Channel.INFO,
+                RequestHandler.RESPONSE,
+                {
+                    shortcuts: Shortcut.getInstance().get('shortcuts'),
+                    cache: await this.browser.webContents.session.getCacheSize(),
+                    ...Status.getInstance().data,
+                },
             )
         }
     }
 
     private onSwitch(
-        _: Electron.IpcMainEvent,
+        _: IpcMainEvent,
         scene: Scenes,
         address?: string,
         handler?: RequestHandler,
@@ -74,14 +94,12 @@ export abstract class AbsWindowIPC extends AbsWindowMenu {
         }
     }
 
-    private onHistory(
-        _: Electron.IpcMainEvent,
-        handler: RequestHandler,
-        index: number,
-    ) {
+    private onHistory(_: IpcMainEvent, handler: RequestHandler, index: number) {
         switch (handler) {
             case RequestHandler.REQUEST:
-                this.centre.sendHistory(
+                this.centre.webContents.send(
+                    Channel.HISTORY,
+                    RequestHandler.RESPONSE,
                     this.browser.webContents.navigationHistory.getAllEntries(),
                 )
                 return
@@ -98,14 +116,18 @@ export abstract class AbsWindowIPC extends AbsWindowMenu {
     }
 
     private onBookmarks(
-        _: Electron.IpcMainEvent,
+        _: IpcMainEvent,
         handler: RequestHandler,
         bookmark: Bookmark,
         index: number,
     ) {
         switch (handler) {
             case RequestHandler.REQUEST:
-                this.centre.sendBookmarks()
+                this.centre.webContents.send(
+                    Channel.BOOKMARK,
+                    RequestHandler.RESPONSE,
+                    Bookmarks.getInstance().get(),
+                )
                 return
             case RequestHandler.ADD:
                 Bookmarks.getInstance().push(bookmark)
@@ -115,6 +137,50 @@ export abstract class AbsWindowIPC extends AbsWindowMenu {
                 return
             case RequestHandler.REMOVE:
                 Bookmarks.getInstance().remove(bookmark as unknown as number)
+                return
+        }
+    }
+
+    private onAnchors(_: IpcMainEvent, handler: RequestHandler, url: string) {
+        switch (handler) {
+            case RequestHandler.REQUEST:
+                this.centre.webContents.send(
+                    Channel.ANCHOR,
+                    RequestHandler.RESPONSE,
+                    Anchors.getInstance().get(),
+                )
+                return
+
+            case RequestHandler.REMOVE:
+                Anchors.getInstance().remove(url)
+                return
+        }
+    }
+
+    private onPopupBlocker(
+        _: IpcMainEvent,
+        handler: RequestHandler,
+        host: string,
+    ) {
+        switch (handler) {
+            case RequestHandler.REQUEST:
+                const blocked = Popup.getInstance().get('blocked')
+                const allowed = Popup.getInstance().get('allowed')
+                Logger.getInstance().log(
+                    'Popup blocker request: ',
+                    blocked,
+                    allowed,
+                )
+                this.centre.webContents.send(
+                    Channel.POPUP_BLOCKER,
+                    RequestHandler.RESPONSE,
+                    Array.from(blocked as string[]),
+                    Array.from(allowed as string[]),
+                )
+                return
+
+            case RequestHandler.MODIFY:
+                Popup.getInstance().toggle(host)
                 return
         }
     }
